Removed Errors from Latest gfortran
I had a few errors in the code pointed out by the latest gfortran (4.7.0), and
I just wanted to clear them up. I don't for a second think they mattered, as
the vast majority was in the getdk.F, and that's just parsing the deck. But
it's nice to make warning-free code.